refactor: reduce duplicated code and clean
