+++ service/src/common/address.rs
@@ -1,7 +1,9 @@
 // Copyright 2023-, GraphOps and Semiotic Labs.
 // SPDX-License-Identifier: Apache-2.0
 
-use ethers::signers::{coins_bip39::English, LocalWallet, MnemonicBuilder, Wallet, WalletError};
+use ethers::signers::{
+    coins_bip39::English, LocalWallet, MnemonicBuilder, Signer, Wallet, WalletError,
+};
 use ethers_core::{k256::ecdsa::SigningKey, utils::hex};
 use sha3::{Digest, Keccak256};
 
@@ -20,4 +22,9 @@ pub fn build_wallet(value: &str) -> Result<Wallet<SigningKey>, WalletError> {
     value
         .parse::<LocalWallet>()
         .or(MnemonicBuilder::<English>::default().phrase(value).build())
+}
+
+/// Get wallet public address to String
+pub fn wallet_address(wallet: &Wallet<SigningKey>) -> String {
+    format!("{:?}", wallet.address())
 }
